# api/protobuf/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/protobuf/api.proto",
    "version": "version not set"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/call_back/transaction/notify": {
      "post": {
        "operationId": "SinohopeService_TransactionNotify",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protobufTransactionNotifyResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protobufTransactionNotifyRequest"
            }
          }
        ],
        "tags": [
          "SinohopeService"
        ]
      }
    },
    "/v1/call_back/withdrawal/confirm": {
      "post": {
        "operationId": "SinohopeService_WithdrawalConfirm",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protobufWithdrawalConfirmResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protobufWithdrawalConfirmRequest"
            }
          }
        ],
        "tags": [
          "SinohopeService"
        ]
      }
    },
    "/v1/check": {
      "post": {
        "operationId": "MpcService_MpcCheck",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protobufMpcCheckResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protobufMpcCheckRequest"
            }
          }
        ],
        "tags": [
          "MpcService"
        ]
      }
    },
    "/v1/hello": {
      "get": {
        "operationId": "HelloService_GetHello",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protobufHelloResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "tags": [
          "HelloService"
        ]
      }
    }
  },
  "definitions": {
    "HelloResponseData": {
      "type": "object",
      "properties": {
        "info": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "type_url": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "protobufHelloResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "format": "int64"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/HelloResponseData"
        }
      }
    },
    "protobufMpcCheckRequest": {
      "type": "object",
      "properties": {
        "callback_id": {
          "type": "string"
        },
        "request_type": {
          "type": "string"
        },
        "request_detail": {
          "type": "object"
        },
        "extra_info": {
          "type": "object"
        }
      }
    },
    "protobufMpcCheckResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "error": {
          "type": "string"
        },
        "data": {
          "type": "object"
        },
        "signature": {
          "type": "string"
        }
      }
    },
    "protobufNullValue": {
      "type": "string",
      "enum": [
        "NULL_VALUE"
      ],
      "default": "NULL_VALUE",
      "description": "`NullValue` is a singleton enumeration to represent the null value for the\n`Value` type union.\n\nThe JSON representation for `NullValue` is JSON `null`.\n\n - NULL_VALUE: Null value."
    },
    "protobufTransactionNotifyRequest": {
      "type": "object",
      "properties": {
        "requestType": {
          "type": "string",
          "format": "int64"
        },
        "requestId": {
          "type": "string"
        },
        "requestDetail": {
          "type": "object"
        },
        "extraInfo": {
          "type": "string"
        }
      }
    },
    "protobufTransactionNotifyResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "format": "int64"
        },
        "requestId": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "protobufWithdrawalConfirmRequest": {
      "type": "object",
      "properties": {
        "requestId": {
          "type": "string"
        },
        "requestDetail": {
          "type": "object"
        },
        "extraInfo": {
          "type": "string"
        }
      }
    },
    "protobufWithdrawalConfirmResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "format": "int64"
        },
        "requestId": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "action": {
          "type": "string"
        }
      }
    },
    "runtimeError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}